+++ src/models/getContributors.go
@@ -32,7 +32,7 @@ type ContributorData []struct {
 }
 
 //获取项目贡献者信息的接口，返回string,不同contributor在字符串中用" "分割
-func getContributors(userName string, programName string) string {
+func GetContributors(userName string, programName string) string {
 	var url_1 string = "https://api.github.com/repos/" + userName + "/" + programName + "/" + "contributors"
 
 	client := &http.Client{}
@@ -62,7 +62,7 @@ func GetContributorNum(url string) int {
 	user_Name := str2[0]
 	program_Name := str2[1]
 
-	info := getContributors(user_Name, program_Name)
+	info := GetContributors(user_Name, program_Name)
 	res := strings.TrimSpace(info)
 	str_arr := strings.Split(res, " ")
 	count := len(str_arr)
